MB-33174: Match param type of setMutationMemoryThreshold to config
1) Remove the cumbersome requirement that callers of
setMutationMemoryThreshold must pass a 0.0 to 1.0 value.

2) Throw if out-of-range values are attempted rather than silently
ignoring.

Reviewed-on: http://review.couchbase.org/105385
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/vbucket.h
@@ -1424,9 +1424,10 @@ public:
      * Set the memory threshold on the current bucket quota for accepting a
      * new mutation. This is same across all the vbuckets
      *
-     * @param memThreshold Threshold between 0 and 1
+     * @param memThreshold Threshold between 0 and 100, value is converted to
+     *        double and divided by 100, e.g. 90 sets a value of 0.9
      */
-    static void setMutationMemoryThreshold(double memThreshold);
+    static void setMutationMemoryThreshold(size_t memThreshold);
 
     /**
      * Check if this StoredValue has become logically non-existent.
